Stop storing indentation in mapped source.
Instead work out the relevant position by finding the start of content
and working from there. It doesn't quite get it right if the position is
in the middle of the indentation, but that's fine.

I'm doing this because I want to be able to change the indentation if
it's a mix of tabs and spaces (which it currently messes up).

git-svn-id: http://svn.boost.org/svn/boost/trunk@84148 b8fc166d-592f-0410-95f2-cb63ce0dd405